file structure for mpt model was refactored
